refactor: rename LessonController.js
