+++ src/routes/LessonRouter.js
@@ -1,5 +1,5 @@
 import express from "express";
-import LessonController from "../controllers/LessonsController.js";
+import LessonController from "../controllers/LessonController.js";
 const router = express.Router();
 
 router.post("/", LessonController.createLesson);
